added log to file to subghz frequency analyzer
Added the ability to save the captured frequency and rssi to a log file.
This can be enabled or disabled by renaming a file in the subghz folder: if a file named fqalogon.txt is found the logging to file is enabled. If file is renamed to fqalogoff.txt logging is disabled. By using a filename as a flag you can enable and disable logging directly on the flipper with the file browser.
At the first startup of Frequency Analyzer it would create (if necessary) a default file named fqalogoff.txt (by default logging is disabled) with some instructions inside, use the internal browser to read them.
The logging is done in files with the capture date used in the name, so every day a new file will be created. The file is also constructed so that it can be easily imported into your preferred spreadsheet for subsequent analysis. The log file is basically a CSV but with a TXT extension, with this trick it can also be opened directly on the flipper with the internal browser for a fast check if you need to.